+++ www/Site/Library/layouts/Admin/Bookmarks/item_view/item_view.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * Виджет пунктов меню
+ * Для отображения свойств объекта соответствующими виджетом из списка
+ * @version 1.0
+ */
+namespace Library\layouts\Admin\Bookmarks\item_view;
+
+use Library\menus\Menu\item_view\item_view as item_view_1;
+
+class item_view extends item_view_1
+{
+    protected $_cut_contents_url = false;
+}